+++ src/json/mod.rs
@@ -86,7 +86,7 @@
 //!
 //! **Happy encoding!**
 
-pub trait Encode: Sync + Send {
+pub trait Encode {
     fn encode(&self, buf: &mut Vec<u8>);
 }