+++ src/number/mod.rs
@@ -39,7 +39,7 @@
 //!
 //! **Happy encoding!**
 
-pub trait Encode: Sync + Send {
+pub trait Encode {
     fn encode(&self, buf: &mut Vec<u8>);
 }