+++ src/json/mod.rs
@@ -88,6 +88,13 @@
 
 pub trait Encode {
     fn encode(&self, buf: &mut Vec<u8>);
+
+    #[inline]
+    fn stringify(&self) -> String {
+        let mut buf = vec![];
+        self.encode(&mut buf);
+        unsafe { String::from_utf8_unchecked(buf) }
+    }
 }
 
 mod array;