+++ src/json/option.rs
@@ -17,5 +17,6 @@ fn test() {
     opt.encode(&mut buf);
     opt = None;
     opt.encode(&mut buf);
+    assert_eq!(opt.stringify(), r#"null"#);
     assert_eq!(String::from_utf8_lossy(&buf), r#"1null"#);
 }