+++ HISTORY.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### Changed
+
+- remove sync + send
+
 ## [0.2.0] - 2024-01-20
 
 ### Added